ai-engine/requirements.txt
# SentinelAI AI Engine Dependencies
# ==================================
# Install with: pip install -r requirements.txt

# MongoDB connection
pymongo>=4.6.1

# Data processing
pandas>=2.1.4
numpy>=1.26.3

# Environment variable management
python-dotenv>=1.0.0

# Text vectorization
scikit-learn>=1.3.2

# Deep learning (Autoencoder)
tensorflow>=2.16.0

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
